* commit '13aa1d1b0569f21304545c005c5da9f810a41b66': (41 commits)
  Continue to allow `system:nodes` to act as `system:node-proxier`
  Rename cluster, user, context to 'default'
  Use localhost instead of 127.0.0.1
  Add comment
  Use service account token in master kube-proxies
  Address review points
  Convert kube-proxy to a DaemonSet
  Add in missing ASG IAM permission for experimental.nodeDrainer.
  Support deployment to AWS GovCloud
  Update kubedns
  Only add CIDR from NLB-backed endpoints to controller SG
  Fixes #946 Support drop-ins in customSystemdUnits
  Remove tests for constraints that are no longer true
  Add support for apiAccessAllowedSourceCIDRs to NLB-backed endpoints
  Use correct predicate function
  Fix logic to check whether the load balancer is a NLB
  Add support for NLB-backed API endpoints
  Bump default k8s to 1.7.6
  Fixes #942 etcdadm cannot find ca.pem
  Remove `createRecordSet` to help newcomers bring up their first clusters (#936)
  ...

+++ test/helper/helper.go
@@ -4,7 +4,6 @@ import (
 	"fmt"
 	"io/ioutil"
 	"os"
-	"path/filepath"
 )
 
 func WithTempDir(fn func(dir string)) {
@@ -20,6 +19,14 @@ func WithTempDir(fn func(dir string)) {
 }
 
 func WithDummyCredentials(fn func(dir string)) {
+	withDummyCredentials(true, fn)
+}
+
+func WithDummyCredentialsButCAKey(fn func(dir string)) {
+	withDummyCredentials(false, fn)
+}
+
+func withDummyCredentials(alsoWriteCAKey bool, fn func(dir string)) {
 	dir, err := ioutil.TempDir("", "dummy-credentials")
 
 	if err != nil {
@@ -38,24 +45,41 @@ func WithDummyCredentials(fn func(dir string)) {
 		}
 		defer os.Remove(certFile)
 
-		keyFile := fmt.Sprintf("%s/%s-key.pem", dir, pairName)
-		if err := ioutil.WriteFile(keyFile, []byte("dummykey"), 0644); err != nil {
-			panic(err)
+		if pairName != "ca" || alsoWriteCAKey {
+			keyFile := fmt.Sprintf("%s/%s-key.pem", dir, pairName)
+			if err := ioutil.WriteFile(keyFile, []byte("dummykey"), 0644); err != nil {
+				panic(err)
+			}
+			defer os.Remove(keyFile)
 		}
-		defer os.Remove(keyFile)
 	}
 
-	symlinks := []struct {
+	type symlink struct {
 		from string
 		to   string
-	}{
+	}
+
+	symlinks := []symlink{
 		{"ca.pem", "worker-ca.pem"},
 		{"ca.pem", "etcd-trusted-ca.pem"},
-		{"ca-key.pem", "worker-ca-key.pem"},
+	}
+
+	if alsoWriteCAKey {
+		symlinks = append(symlinks, symlink{"ca-key.pem", "worker-ca-key.pem"})
+	}
+
+	wd, err := os.Getwd()
+	if err != nil {
+		panic(err)
+	}
+
+	if err := os.Chdir(dir); err != nil {
+		panic(err)
 	}
 
 	for _, sl := range symlinks {
-		to := filepath.Join(dir, sl.to)
+		from := sl.from
+		to := sl.to
 
 		if _, err := os.Lstat(to); err == nil {
 			if err := os.Remove(to); err != nil {
@@ -63,10 +87,14 @@ func WithDummyCredentials(fn func(dir string)) {
 			}
 		}
 
-		if err := os.Symlink(sl.from, to); err != nil {
+		if err := os.Symlink(from, to); err != nil {
 			panic(err)
 		}
 		defer os.Remove(to)
+	}
+
+	if err := os.Chdir(wd); err != nil {
+		panic(err)
 	}
 
 	fn(dir)
